# platform/swaggo/swagger.yaml
basePath: /flighthours/api/v1
definitions:
...
  handlers.ChangePasswordRequest:
    properties:
      confirm_password:
        minLength: 8
        type: string
      current_password:
        type: string
      email:
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - confirm_password
    - current_password
    - email
    - new_password
    type: object
  handlers.ChangePasswordResponse:
    properties:
      changed:
        type: boolean
      email:
        type: string
    type: object
...
  handlers.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
  handlers.PasswordResetRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  handlers.PasswordResetResponse:
    properties:
      sent:
        type: boolean
    type: object
  handlers.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handlers.ResendVerificationEmailRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  handlers.ResendVerificationEmailResponse:
    properties:
      email:
        type: string
      sent:
        type: boolean
    type: object
...
  handlers.UpdatePasswordResponse:
    properties:
      email:
        type: string
      updated:
        type: boolean
    type: object
...
  handlers.VerifyEmailRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  middleware.ErrorResponse:
    properties:
      code:
        type: string
      fields:
        items:
          $ref: '#/definitions/middleware.ValidationFieldError'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@champion19.com
    name: Champion19 Support
  description: API para gestión de horas de vuelo de pilotos y tripulantes
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Flighthours API
  version: "1.0"
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Allows an authenticated user to change their password by providing
        the current password
      parameters:
      - description: Current and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed
          schema:
            $ref: '#/definitions/handlers.ChangePasswordResponse'
        "400":
          description: Passwords don't match
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Current password incorrect
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change authenticated user password
      tags:
      - Authentication
  /auth/password-reset:
    post:
      consumes:
      - application/json
      description: Sends an email with instructions to reset the password. Always
        returns success for security.
      parameters:
      - description: Email for password reset
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reset email sent
          schema:
            $ref: '#/definitions/handlers.PasswordResetResponse'
        "400":
          description: Invalid email
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Request password reset
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the access token using a valid refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New tokens
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Refresh access token
      tags:
      - Authentication
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resends the verification email to a registered user who hasn't
        verified their account
      parameters:
      - description: Email to resend verification
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResendVerificationEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email sent
          schema:
            $ref: '#/definitions/handlers.ResendVerificationEmailResponse'
        "400":
          description: Invalid email
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Email already verified
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Resend verification email
      tags:
      - Authentication
  /auth/update-password:
    post:
      consumes:
      - application/json
      description: Allows updating the password using a token received by email. Token
        expires in 12 hours.
      parameters:
      - description: Token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated
          schema:
            $ref: '#/definitions/handlers.UpdatePasswordResponse'
        "400":
          description: Passwords don't match or invalid token
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Expired or invalid token
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Update password with reset token
      tags:
      - Authentication
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verifies the user's email using a JWT proxy token
      parameters:
      - description: Verification token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email verified
          schema:
            $ref: '#/definitions/handlers.VerifyEmailResponse'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Email already verified
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Verify user email
      tags:
      - Authentication
...
  /employees:
    delete:
      description: Permanently deletes the authenticated employee's account from the
        system (Keycloak + DB)
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Employee not found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete authenticated user account
      tags:
      - Employees
    get:
      description: Returns the profile information of the authenticated employee
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/handlers.EmployeeResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get authenticated user profile
      tags:
      - Employees
    put:
      consumes:
      - application/json
      description: 'Updates the basic information of the authenticated employee. Preserves:
        email, password, airline, bp.'
      parameters:
      - description: Employee data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Employee updated
          schema:
            $ref: '#/definitions/handlers.UpdateEmployeeResponse'
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Employee not found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update authenticated user information
      tags:
      - Employees
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns access tokens
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Invalid credentials format
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Email not verified or incorrect credentials
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: User login
      tags:
      - Authentication
...
  /register:
    post:
      consumes:
      - application/json
      description: Registers a new employee in the system and sends a verification
        email
      parameters:
      - description: Employee registration data
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/handlers.EmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Employee registered successfully
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Email or ID already exists
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Register a new employee
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token (format: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"